% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ui.R
\name{add_checkmark}
\alias{add_checkmark}
\title{Add a green checkmark to a div}
\usage{
add_checkmark(ns = NULL, id = "quiz-container", element = "h3")
}
\arguments{
\item{ns}{namespace of the Shiny module}

\item{id}{id of the div to add the checkmark to}

\item{element}{element within the div to place the checkmark}
}
\value{
called for side effect
}
\description{
Add a green checkmark to a div
}
\author{
Joseph Marlo
}
\keyword{internal}
