% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ui.R
\name{add_confetti}
\alias{add_confetti}
\title{Add confetti celebration animation}
\usage{
add_confetti()
}
\value{
div containing divs of class "confetti-piece"
}
\description{
Requires confetti.css to be in the www/css folder
}
\examples{
add_confetti()
}
\seealso{
\link{https://codepen.io/zer0kool/pen/KjZWRW}
}
\author{
Joseph Marlo
}
