% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-module.R
\name{add_external_resources}
\alias{add_external_resources}
\title{External resources to include in the app}
\usage{
add_external_resources()
}
\value{
an invisible object of class shiny.tag
}
\description{
Examples include \code{shinyjs::useShinyjs} or \code{fontawesome::fa_html_dependency}. These objects are raised to the head of html document.
}
\author{
Joseph Marlo
}
