% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ui.R
\name{add_message_correct}
\alias{add_message_correct}
\alias{add_message_wrong}
\alias{add_message_skipped}
\title{Add ending messages to the quiz}
\usage{
add_message_correct(text)

add_message_wrong(text)

add_message_skipped(text)
}
\arguments{
\item{text}{Message to display}
}
\value{
called for side effect
}
\description{
Add ending messages to the quiz
}
\section{Functions}{
\itemize{
\item \code{add_message_correct()}: Message to display when user is correct

\item \code{add_message_wrong()}: Message to display when user is wrong

\item \code{add_message_skipped()}: Message to display when user skipped the quiz

}}
\examples{
#TBD
}
\author{
Joseph Marlo
}
\keyword{internal}
