% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ui.R
\name{add_progress_bar}
\alias{add_progress_bar}
\title{Create a bootstrap progress bar}
\usage{
add_progress_bar(percent, bg_color)
}
\arguments{
\item{percent}{The percent to fill the progress bar to}

\item{bg_color}{Background color}
}
\value{
html
}
\description{
Create a bootstrap progress bar
}
\seealso{
\url{https://getbootstrap.com/docs/4.0/components/progress/}
}
\author{
Joseph Marlo
}
\keyword{internal}
