% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ui.R
\name{add_red_x}
\alias{add_red_x}
\title{Add a red X to a div}
\usage{
add_red_x(ns = NULL, id = "quiz-container", element = "h3")
}
\arguments{
\item{ns}{namespace of the Shiny module}

\item{id}{id of the div to add the X to}

\item{element}{element within the div to place the X}
}
\value{
called for side effect
}
\description{
Add a red X to a div
}
\author{
Joseph Marlo
}
\keyword{internal}
