% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{construct_quiz}
\alias{construct_quiz}
\alias{construct_question}
\title{Constructors for the quiz}
\usage{
construct_quiz(..., options = set_quiz_options())

construct_question(prompt, answerUserDisplay, answerCorrectDisplay, grader)
}
\arguments{
\item{...}{objects of class 'quizQuestions'. See \code{\link[=construct_question]{construct_question()}}}

\item{options}{a list of options generated from \code{\link[=set_quiz_options]{set_quiz_options()}}}

\item{prompt}{an \link[htmltools:builder]{htmltools::div} that represents a quiz question}

\item{answerUserDisplay}{a function that takes the user answer and prints it neatly. This is wrapped with \code{\link[purrr:possibly]{purrr::possibly()}} to catch any errors.}

\item{answerCorrectDisplay}{a character that prints the correct answer neatly}

\item{grader}{a function that takes the user answer and determines if it is correct. Must return TRUE or FALSE. This is wrapped with \code{\link[purrr:possibly]{purrr::possibly()}} to catch any errors.}
}
\value{
an object of class \code{quiz}

an object of class \code{quizQuestion}
}
\description{
Construct a quiz or quiz question
}
\details{
See dev/example-app.R for current example.

TODO: This should probably used only on the backend and be internal only?
}
\section{Functions}{
\itemize{
\item \code{construct_quiz()}: Construct a quiz object

\item \code{construct_question()}: Construct a question object

}}
\examples{
#TBD
}
\seealso{
\code{\link[=construct_question]{construct_question()}}, \code{\link[=set_quiz_options]{set_quiz_options()}}, \code{\link[=construct_messages]{construct_messages()}}
}
\author{
Joseph Marlo
}
