% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creators.R
\name{create_question}
\alias{create_question}
\title{Create a quiz question}
\usage{
create_question(
  prompt,
  ...,
  type = c("auto", "single", "multiple"),
  input = c("auto", "select", "checkbox"),
  id = "quiz",
  id_parent = character(0)
)
}
\arguments{
\item{prompt}{Text of the question prompt}

\item{...}{Objects of class 'quizChoice' generated from \code{\link[=add_choice]{add_choice()}} or \code{\link[=add_slider]{add_slider()}}. Named options to \link[shiny:selectInput]{shiny::selectInput} or \link[shiny:checkboxGroupInput]{shiny::checkboxGroupInput} can be passed as well.}

\item{type}{One of c('auto', 'single', 'multiple'). How many answers are allowed}

\item{input}{One of c('auto', 'select', 'checkbox')}

\item{id}{Namespace of the module. Defaults to 'quiz'. This only needs to modified if there are multiple quizzes.}

\item{id_parent}{Namespace of the parent module, if any. This only needs to modified if the quiz sits within a shiny module.}
}
\value{
an object of class 'quizQuestion'
}
\description{
Create a quiz question
}
\examples{
\dontrun{
q <- create_question(
  prompt = 'My prompt explaining what the ATE of this thing should be',
  add_choice("34"),
  add_choice("59", TRUE),
  add_choice("98", TRUE)
)
preview_question(q)

q2 <- create_question(
  prompt = 'My prompt explaining what the ATC of this thing should be',
  add_slider(0, 30, 15, correct = 10)
 )
create_quiz(q, q2)
}
}
\seealso{
\code{\link[=add_choice]{add_choice()}}
}
\author{
Joseph Marlo, George Perrett
}
