% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creators.R
\name{create_question_raw}
\alias{create_question_raw}
\title{Create a quiz question}
\usage{
create_question_raw(
  prompt,
  grader,
  correct_answer_pretty,
  user_answer_prettifier = function(user_input) paste0(user_input, collapse = ", ")
)
}
\arguments{
\item{prompt}{Text of the question prompt. Preferably wrapped in \code{\link[htmltools:builder]{htmltools::div()}}.}

\item{grader}{A function with one argument that takes the user input and returns TRUE/FALSE}

\item{correct_answer_pretty}{A string representing the correct answer that is printed 'pretty'}

\item{user_answer_prettifier}{A function with one argument that takes the user's answers and prints it 'pretty'}
}
\value{
a object of class \code{quizQuestion}
}
\description{
Create a quiz question using custom inputs. This is a more flexible function that \code{\link[=create_question]{create_question()}}. It allows any html in the \code{prompt}. ADD REQUIREMENTS
}
\examples{
\dontrun{
q <- create_question_raw(
  htmltools::div(
    htmltools::p("my question"),
    shiny::selectInput(
      inputId = shiny::NS('quiz')('answers'),
      label = 'Select 5',
      choices = c(4, 5, 6)
    )
 ),
 grader = \(user_input) user_input == '5',
 correct_answer_pretty = '5'
)
preview_question(q)
}
}
\author{
Joseph Marlo
}
