% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state-machine.R
\name{format_prompts}
\alias{format_prompts}
\alias{quiz_format_prompt}
\title{Add headers with question numbers to the quiz questions}
\usage{
format_prompts(quiz)

quiz_format_prompt(prompt, i)
}
\arguments{
\item{quiz}{a quiz; see ...?}
}
\value{
quiz
}
\description{
Add headers with question numbers to the quiz questions
}
\section{Functions}{
\itemize{
\item \code{quiz_format_prompt()}: Add a header denoting the question number

}}
\examples{
#TBD
}
\author{
Joseph Marlo
}
