% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state-machine.R
\name{format_questions}
\alias{format_questions}
\alias{quiz_format_question_text}
\title{Add headers with question numbers to the quiz questions}
\usage{
format_questions(quiz)

quiz_format_question_text(question, i)
}
\arguments{
\item{quiz}{a quiz; see ...?}
}
\value{
quiz
}
\description{
Add headers with question numbers to the quiz questions
}
\section{Functions}{
\itemize{
\item \code{quiz_format_question_text()}: Add a header denoting the question number

}}
\examples{
#TBD
}
\author{
Joseph Marlo
}
