% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%notin\%}
\alias{\%notin\%}
\title{Check to see if value is not in a vector}
\usage{
... \%notin\% NA
}
\arguments{
\item{...}{}
}
\value{
boolean
}
\description{
This is just a negative of \code{\%in\%}
}
\examples{
1 \%notin\% 2:4
}
\author{
Joseph Marlo
}
