% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_truthy}
\alias{is_truthy}
\title{Check if a value is truthy}
\usage{
is_truthy(x)
}
\arguments{
\item{x}{An expression whose truthiness value we want to determine}
}
\value{
boolean
}
\description{
A value is truthy unless it is FALSE, NA, NULL, an empty data.frame, or empty list.
}
\details{
Modified from shiny::isTruthy
}
\examples{
is_truthy(TRUE)
is_truthy(FALSE)
is_truthy(1)
is_truthy(0)
is_truthy(NULL)
is_truthy(NA)
is_truthy(data.frame())
is_truthy(data.frame(x = 1))
}
\seealso{
\code{\link[shiny:isTruthy]{shiny::isTruthy()}}
}
\author{
Joseph Marlo
}
