% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-module.R
\name{mod_quiz_ui}
\alias{mod_quiz_ui}
\alias{mod_quiz_server}
\title{quiz module}
\usage{
mod_quiz_ui(id)

mod_quiz_server(
  id,
  id_parent = character(0),
  questions,
  message_correct,
  message_wrong,
  message_skipped,
  embed_quiz = TRUE,
  sandbox_mode = FALSE
)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}
}
\description{
A shiny Module to implement the quiz
}
\section{Functions}{
\itemize{
\item \code{mod_quiz_server()}: Server side function

}}
\author{
Joseph Marlo
}
