% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview-tools.R
\name{preview_question}
\alias{preview_question}
\title{Preview a question in the viewer}
\usage{
preview_question(question)
}
\arguments{
\item{question}{question to preview}
}
\value{
Called for side effect
}
\description{
Preview a question in the viewer
}
\examples{
# TBD
}
\author{
Joseph Marlo
}
