% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview-tools.R
\name{preview_quiz}
\alias{preview_quiz}
\alias{preview_question}
\title{Tools for previewing quizzes}
\usage{
preview_quiz(quiz)

preview_question(question)
}
\arguments{
\item{quiz}{an object of class 'quiz' to preview}

\item{question}{an object of class 'quizQuestion' to preview}
}
\value{
Called for side effect
}
\description{
Launch a viewer to preview the html structure of the questions in a quiz.
}
\section{Functions}{
\itemize{
\item \code{preview_quiz()}: Preview a quiz

\item \code{preview_question()}: Preview a single question

}}
\examples{
# TBD
}
\author{
Joseph Marlo
}
