% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\docType{class}
\name{quiz-class}
\alias{quiz-class}
\title{S4 class for a quiz}
\value{
none, sets a class
}
\description{
S4 class for a quiz
}
\section{Slots}{

\describe{
\item{\code{questions}}{list. A list of \code{quizQuestion}s}

\item{\code{options}}{TBD.}
}}

\seealso{
\code{\link[=construct_quiz]{construct_quiz()}}
}
\author{
Joseph Marlo
}
\keyword{internal}
