% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiz_status_functions.R
\name{quiz_complete}
\alias{quiz_complete}
\title{Returns TRUE or FALSE if the user has answered all the quiz questions.}
\usage{
quiz_complete(mod)
}
\arguments{
\item{mod}{the ShinyQuiz module}
}
\value{
TRUE if the user has viewed all the quiz questions.
}
\description{
Returns TRUE or FALSE if the user has answered all the quiz questions.
}
