% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiz_status_functions.R
\name{quiz_viewed}
\alias{quiz_viewed}
\title{Returns TRUE if the user has viewed all the quiz questions.}
\usage{
quiz_viewed(mod)
}
\arguments{
\item{mod}{the ShinyQuiz module}
}
\value{
TRUE if the user has viewed all the quiz questions.
}
\description{
Returns TRUE if the user has viewed all the quiz questions.
}
