% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quizButton.R
\name{quizButton}
\alias{quizButton}
\title{Button to begin a Shiny Quiz.}
\usage{
quizButton(id, label = "Start Quiz")
}
\arguments{
\item{id}{The id string to be namespaced .}

\item{label}{Label for the button.}
}
\description{
Button to begin a Shiny Quiz.
}
