% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creators.R
\docType{class}
\name{quizChoice-class}
\alias{quizChoice-class}
\title{S4 class for a quiz choices}
\value{
none, sets a class
}
\description{
S4 class for a quiz choices
}
\section{Slots}{

\describe{
\item{\code{text}}{character.}

\item{\code{correct}}{logical.}
}}

\author{
Joseph Marlo
}
