% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creators.R
\docType{class}
\name{quizChoiceNumeric-class}
\alias{quizChoiceNumeric-class}
\title{S4 class for a quiz numeric response}
\value{
none, sets a class
}
\description{
S4 class for a quiz numeric response
}
\section{Slots}{

\describe{
\item{\code{correct}}{numeric.}
}}

\author{
George Perrett
}
\keyword{internal}
