% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creators.R
\docType{class}
\name{quizChoiceSlider-class}
\alias{quizChoiceSlider-class}
\title{S4 class for a quiz sliders}
\value{
none, sets a class
}
\description{
S4 class for a quiz sliders
}
\section{Slots}{

\describe{
\item{\code{min}}{numeric.}

\item{\code{max}}{numeric.}

\item{\code{default}}{numeric.}

\item{\code{correct}}{numeric.}
}}

\author{
Joseph Marlo
}
\keyword{internal}
