% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creators.R
\docType{class}
\name{quizChoiceText-class}
\alias{quizChoiceText-class}
\title{S4 class for a quiz free form text response}
\value{
none, sets a class
}
\description{
S4 class for a quiz free form text response
}
\section{Slots}{

\describe{
\item{\code{correct}}{character}
}}

\author{
Joseph Marlo
}
\keyword{internal}
