% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\docType{class}
\name{quizMessages-class}
\alias{quizMessages-class}
\title{S4 class for a quiz messages to display at the end}
\value{
none, sets a class
}
\description{
S4 class for a quiz messages to display at the end
}
\section{Slots}{

\describe{
\item{\code{message_correct}}{character.}

\item{\code{message_wrong}}{character.}

\item{\code{message_skipped}}{character.}
}}

\seealso{
\code{\link[=construct_messages]{construct_messages()}}
}
\author{
Joseph Marlo
}
