% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\docType{class}
\name{quizQuestion-class}
\alias{quizQuestion-class}
\title{S4 class for a quiz question}
\value{
none, sets a class
}
\description{
S4 class for a quiz question
}
\section{Slots}{

\describe{
\item{\code{prompt}}{shiny.tag.}

\item{\code{answerUser}}{list.}

\item{\code{answerUserPrettifier}}{function.}

\item{\code{answerCorrectPretty}}{character.}

\item{\code{grader}}{function.}
}}

\seealso{
\code{\link[=construct_question]{construct_question()}}
}
\author{
Joseph Marlo
}
\keyword{internal}
