% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiz_module_server.R
\name{quizServer}
\alias{quizServer}
\title{Shiny Quiz Server Module}
\usage{
quizServer(
  id,
  quiz,
  title = "Shiny Quiz",
  show_on_startup = TRUE,
  size = "xl",
  fade = TRUE,
  next_button = "Next",
  previous_button = "Previous",
  finish_button = "Finish",
  cancel_button = NULL,
  progress_bar = TRUE,
  required_message = "You need to answer this question before moving on."
)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's UI function.}

\item{quiz}{a list with the quiz details.}

\item{title}{the title of the modal dialog.}

\item{show_on_startup}{if TRUE the quiz will be displayed when the Shiny app starts.}

\item{size}{One of "s" for small, "m" (the default) for medium, or "l" for large.}

\item{fade}{If FALSE, the modal dialog will have no fade-in animation
(it will simply appear rather than fade in to view).}

\item{progress_bar}{whether to display a progress bar.}

\item{required_message}{the message displayed to the user when they don't anser
a required message.}
}
\description{
\if{html}{\out{<div class="sourceCode">}}\preformatted{quiz <- list(
    directions = '',
    questions = list(
        list(
            stem = 'Question 1',
            type = 'multiple_choice',
            choices = c(''),
            required = FALSE,
            multiple = FALSE
        ),
        list(
            stem = 'Question 2',
            type = 'text',
            required = FALSE
        )
    )
)
}\if{html}{\out{</div>}}
}
\details{
Types include:
\itemize{
\item \code{multiple_choice} - Multiple choice questions using  \code{selectInput}.
\item \code{text} - Text input using \code{textAreaInput}.
\item \code{numeric} - Numeric input using \code{numericInput}.
\item \code{slider} - Numeric input using \code{sliderInput}.
}
}
