% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state-machine.R
\name{resample_questions_if_sandbox}
\alias{resample_questions_if_sandbox}
\title{Create quasi infinite quiz by resampling questions n times}
\usage{
resample_questions_if_sandbox(quiz, sandbox_mode, n = 50)
}
\arguments{
\item{quiz}{A list of questions of class 'quizQuestion'}

\item{sandbox_mode}{boolean}

\item{n}{Number of resamples to make}
}
\value{
questions
}
\description{
Create quasi infinite quiz by resampling questions n times
}
\examples{
#TBD
}
\author{
Joseph Marlo
}
