% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ui.R
\name{scroll_to_div}
\alias{scroll_to_div}
\title{Scroll the webpage to a certain div}
\usage{
scroll_to_div(ns = NULL, id = "quiz-container")
}
\arguments{
\item{ns}{namespace of the Shiny module}

\item{id}{id of the div}
}
\value{
called for side effect
}
\description{
Scroll the webpage to a certain div
}
\author{
Joseph Marlo
}
\keyword{internal}
