% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{set_quiz_options}
\alias{set_quiz_options}
\title{Set the options for the quiz}
\usage{
set_quiz_options(messages, sandbox = FALSE, embed = FALSE, ...)
}
\arguments{
\item{messages}{an object of class \code{quizMessages} containing the messages to show at the end. If not provided, defaults are used.}

\item{sandbox}{boolean. TBD}

\item{embed}{boolean. TBD TODO: remove?}

\item{...}{other named options to pass to \code{quiz}}
}
\value{
a list
}
\description{
These are options to be passed to a \code{quiz}.
}
\seealso{
\code{\link[=construct_quiz]{construct_quiz()}}
}
