% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyQuiz-package.R
\docType{package}
\name{ShinyQuiz-package}
\alias{ShinyQuiz}
\alias{ShinyQuiz-package}
\title{ShinyQuiz: Framework for Collecting Data from Users in Shiny Applications}
\description{
ShinyQuiz provides a framework for collecting data from users through a series of questions. Common uses include short multiple choice quizzes, survey questions, or essays. The package supports responses as radio buttons (with one selection allowed), check boxes (allowoing for multiple answers per question), numeric responses, and text boxes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jbryer/ShinyQuiz}
  \item Report bugs at \url{https://github.com/jbryer/ShinyQuiz/issues}
}

}
\author{
\strong{Maintainer}: Jason Bryer \email{jason@bryer.org}

}
\keyword{internal}
