% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/quiz.r
\name{shinyQuiz}
\alias{shinyQuiz}
\title{Create a shiny quiz widget}
\usage{
shinyQuiz(id = paste0("quiz_", sample.int(1e+11, 1)), yaml,
  quiz.handler = NULL, add.handler = TRUE, single.check.btn = TRUE,
  defaults = quizDefaults(lang = lang), lang = "en")
}
\arguments{
\item{id}{the id of the quiz}

\item{yaml}{a yaml string that specifies the quize}

\item{quiz.handler}{a function that will be called if the quiz is checked.
The boolean argument solved is TRUE if the quiz was solved
and otherwise FALSE}
}
\description{
Create a shiny quiz widget
}

