% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{verify_quiz_structure}
\alias{verify_quiz_structure}
\title{Verify that a quiz is a quiz}
\usage{
verify_quiz_structure(quiz)
}
\arguments{
\item{quiz}{a quiz; see ...? Effectively a list of questions}
}
\value{
invisible TRUE if no errors
}
\description{
Verify that a quiz is a quiz
}
\examples{
#TBD
}
\author{
Joseph Marlo
}
