% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creators.R
\name{add_choice}
\alias{add_choice}
\alias{add_slider}
\title{Create a choice for a question}
\usage{
add_choice(text, correct = FALSE)

add_slider(min = 0, max = 1, default_position = 0.5, correct)
}
\arguments{
\item{text}{Text of the choice answer}

\item{correct}{Boolean denoting if this \code{choice} is correct; numeric for \code{slider}}

\item{min}{the minimum value of the slider range}

\item{max}{the maximum value of the slider range}

\item{default_position}{the default value the slider should take}
}
\value{
an object of class 'quizChoice'

an object of class 'quizChoiceSlider'
}
\description{
Create a choice for a question
}
\section{Functions}{
\itemize{
\item \code{add_choice()}: Create a discrete choice

\item \code{add_slider()}: Create a slider choice

}}
\examples{
add_choice('39')
add_choice('39', TRUE)
add_slider(0, 1, 0.5, 0.8)

create_question(
 'My question prompt',
 add_choice('39'),
 add_choice('41', TRUE)
)
}
\seealso{
\code{\link[=create_question]{create_question()}}
}
\author{
Joseph Marlo
}
