% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{construct_quiz}
\alias{construct_quiz}
\alias{construct_question}
\alias{construct_messages}
\title{Constructors for the quiz}
\usage{
construct_quiz(questions, options = set_quiz_options())

construct_question(prompt, answerUserDisplay, answerCorrectDisplay, grader)

construct_messages(message_correct, message_wrong, message_skipped)
}
\arguments{
\item{questions}{a list with objects of class 'quizQuestions'}

\item{prompt}{an \link[htmltools:builder]{htmltools::div} that represents a quiz question}

\item{answerUserDisplay}{a function that takes the user answer and prints it neatly. This is wrapped with \code{\link[purrr:possibly]{purrr::possibly()}} to catch any errors.}

\item{answerCorrectDisplay}{a character that prints the correct answer neatly}

\item{grader}{a function that takes the user answer and determines if it is correct. Must return TRUE or FALSE. This is wrapped with \code{\link[purrr:possibly]{purrr::possibly()}} to catch any errors.}

\item{message_correct}{a string to be shown at the end of the quiz when the user gets all questions correct}

\item{message_wrong}{a string to be shown at the end of the quiz when the user gets at least one question wrong}

\item{message_skipped}{a string to be shown at the end of the quiz when the user skips the quiz or ends it early}

\item{messages}{an object of class 'quizMessages'}
}
\value{
an object of class \code{quiz}

an object of class \code{quizQuestion}

an object of class \code{quizMessages}
}
\description{
Construct a quiz, question, or messages object.
}
\details{
See dev/example-app.R for current example.

TODO: This should probably used only on the backend and be internal only?
}
\section{Functions}{
\itemize{
\item \code{construct_quiz()}: Construct a quiz object

\item \code{construct_question()}: Construct a question object

\item \code{construct_messages()}: Construct a messages object

}}
\examples{
#TBD
}
\seealso{
\code{\link[=set_quiz_options]{set_quiz_options()}}
}
\author{
Joseph Marlo
}
