% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{construct_quiz}
\alias{construct_quiz}
\alias{construct_question}
\alias{construct_messages}
\title{Constructors for the quiz}
\usage{
construct_quiz(questions, messages)

construct_question(prompt, answerUserDisplay, answerCorrectDisplay, grader)

construct_messages(message_correct, message_wrong, message_skipped)
}
\arguments{
\item{prompt}{}

\item{answerUserDisplay}{}

\item{answerCorrectDisplay}{}

\item{grader}{}
}
\value{
TBD
}
\description{
This should probably used only on the backend
}
\section{Functions}{
\itemize{
\item \code{construct_question()}: Construct the question object

\item \code{construct_messages()}: Construct the messages object

}}
\examples{
#TBD
}
\author{
Joseph Marlo
}
