% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creators.R
\name{create_quiz}
\alias{create_quiz}
\title{Create a quiz}
\usage{
create_quiz(..., options = set_quiz_options())
}
\arguments{
\item{...}{objects of class 'quizQuestions'. See \code{\link[=create_question]{create_question()}}, \code{\link[=create_question_raw]{create_question_raw()}}}

\item{options}{a list of options generated from \code{\link[=set_quiz_options]{set_quiz_options()}}}
}
\value{
an object of class \code{quiz}
}
\description{
Create a quiz
}
\examples{
quiz <- create_quiz(
  create_question(
    'Lorem ipsum dolor sit amet, consectetur adipiscing elit. Select nulla.',
    add_choice('auctor'),
    add_choice('nulla', correct = TRUE)
  ),
  create_question(
    'Mauris congue aliquet dui, ut dapibus lorem porttitor sed. Select 600.',
    add_choice('600', correct = TRUE),
    add_choice('800')
  )
)
}
\author{
Joseph Marlo
}
