% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview-tools.R
\name{preview_quiz}
\alias{preview_quiz}
\alias{preview_question}
\title{Tools for previewing quizes}
\usage{
preview_quiz(quiz)

preview_question(question)
}
\arguments{
\item{question}{question to preview}
}
\value{
Called for side effect
}
\description{
Tools for previewing quizes
}
\section{Functions}{
\itemize{
\item \code{preview_question()}: Preview a single question

}}
\examples{
# TBD
}
\author{
Joseph Marlo
}
