% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{class}
\name{quiz-class}
\alias{quiz-class}
\title{S4 class for a quiz}
\value{
none, sets a class
}
\description{
S4 class for a quiz
}
\section{Slots}{

\describe{
\item{\code{questions}}{list. A list of quizQuestions}

\item{\code{messages}}{quizMessages.}
}}

\examples{
# TBD
}
\author{
Joseph Marlo
}
