% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-module.R
\name{quiz_ui}
\alias{quiz_ui}
\alias{quiz_server}
\title{quiz module}
\usage{
quiz_ui(id)

quiz_server(
  id,
  id_parent = character(0),
  quiz,
  embed_quiz = TRUE,
  sandbox_mode = FALSE
)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}

\item{id_parent}{if using within a Shiny module, the id of that module}

\item{quiz}{TBD}

\item{embed_quiz}{boolean. remove?}

\item{sandbox_mode}{boolean. Resample questions for quasi infinite mode?}

\item{id}{a unique string that corresponds that is identical for UI and server}
}
\description{
A shiny Module to implement the quiz
}
\section{Functions}{
\itemize{
\item \code{quiz_server()}: Server side function

}}
\author{
Joseph Marlo
}
