% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-module.R
\name{quiz_ui}
\alias{quiz_ui}
\alias{quiz_server}
\title{Run a quiz in R Shiny}
\usage{
quiz_ui(id)

quiz_server(quiz, id, id_parent = character(0))
}
\arguments{
\item{id}{a unique string that is identical across UI and server}

\item{quiz}{an object of class \code{quiz}. See \code{\link[=construct_quiz]{construct_quiz()}}}

\item{id_parent}{if using within a Shiny module, the id of that module}
}
\value{
a reactive object showing the results of the quiz
}
\description{
A shiny Module to implement a quiz
}
\section{Functions}{
\itemize{
\item \code{quiz_ui()}: UI side function

\item \code{quiz_server()}: Server side function

}}
\seealso{
\code{\link[=create_quiz]{create_quiz()}}
}
\author{
Joseph Marlo
}
