% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-module.R
\name{quiz_ui}
\alias{quiz_ui}
\alias{quiz_server}
\title{Run a quiz in R Shiny}
\usage{
quiz_ui(id)

quiz_server(
  id,
  id_parent = character(0),
  quiz,
  embed_quiz = TRUE,
  sandbox_mode = FALSE
)
}
\arguments{
\item{id}{a unique string that is identical across UI and server}

\item{id_parent}{if using within a Shiny module, the id of that module}

\item{quiz}{TBD}

\item{embed_quiz}{boolean. remove?}

\item{sandbox_mode}{boolean. Resample questions for quasi infinite mode?}
}
\description{
A shiny Module to implement a quiz
}
\section{Functions}{
\itemize{
\item \code{quiz_ui()}: UI side function

\item \code{quiz_server()}: Server side function

}}
\author{
Joseph Marlo
}
