% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{class}
\name{quizQuestion-class}
\alias{quizQuestion-class}
\title{S4 class for a quiz question}
\value{
none, sets a class
}
\description{
S4 class for a quiz question
}
\section{Slots}{

\describe{
\item{\code{question}}{shiny.tag.}

\item{\code{answerUser}}{list.}

\item{\code{answerUserDisplay}}{function.}

\item{\code{answerCorrectDisplay}}{character.}

\item{\code{grader}}{function.}
}}

\examples{
TBD
}
\author{
Joseph Marlo
}
