% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{set_quiz_options}
\alias{set_quiz_options}
\alias{create_messages}
\title{Set the options for the quiz}
\usage{
set_quiz_options(
  ns = shiny::NS("quiz"),
  messages,
  sandbox = FALSE,
  sandbox_resample_n = 50,
  end_on_first_wrong = !sandbox,
  embed = FALSE,
  progress_bar = !sandbox,
  progress_bar_color = "#609963",
  ...
)

create_messages(message_correct, message_wrong, message_skipped)
}
\arguments{
\item{ns}{namespace generated from \code{\link[shiny:NS]{shiny::NS()}}. When using custom namespaces, the individual \code{\link[=create_question]{create_question()}} requires the namespace as well.}

\item{messages}{an object of class \code{quizMessages} generated from \code{\link[=create_messages]{create_messages()}} containing the messages to show at the end. If not provided, defaults are used.}

\item{sandbox}{boolean. TBD}

\item{sandbox_resample_n}{The number of question resamples when in sandbox mode}

\item{embed}{boolean. TBD TODO: remove?}

\item{progress_bar}{boolean. Show the progress bar UI at the top of the quiz}

\item{progress_bar_color}{Color code for the progress bar background}

\item{...}{other named options to pass to \code{quiz}}

\item{message_correct}{a string to be shown at the end of the quiz when the user gets all questions correct}

\item{message_wrong}{a string to be shown at the end of the quiz when the user gets at least one question wrong}

\item{message_skipped}{a string to be shown at the end of the quiz when the user skips the quiz or ends it early}
}
\value{
a list

an object of class \code{quizMessages}
}
\description{
These are options to be passed to a \code{quiz}.
}
\section{Functions}{
\itemize{
\item \code{set_quiz_options()}: Sets the options for a \code{quiz}

\item \code{create_messages()}: Create a messages object

}}
\examples{
\dontrun{
quiz <- create_quiz(..., options = set_quiz_options(sandbox = TRUE))
quiz@options <- set_quiz_options(sandbox = FALSE)
}
}
\seealso{
\code{\link[=construct_quiz]{construct_quiz()}}, \code{\link[=construct_messages]{construct_messages()}}
}
