% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{verify_question_structure}
\alias{verify_question_structure}
\alias{verify_n_args}
\title{Verify a quiz question is the correct format}
\usage{
verify_question_structure(question)

verify_n_args(fn, n)
}
\arguments{
\item{question}{TBD}
}
\value{
invisible TRUE if all tests passed
}
\description{
Verify a quiz question is the correct format
}
\section{Functions}{
\itemize{
\item \code{verify_n_args()}: Verify a function has n arguments

}}
\examples{
# TBD
}
\author{
Joseph Marlo
}
